changed folders and added env variable to build dockerized

+++ stage2/04-automatica-docker/01-run.sh
@@ -1,5 +1,9 @@
 #!/bin/bash -e
 
+if [ -z "$INSTALL_NATIVE" ]
+then
+    exit 0
+fi
 
 rm -f "${ROOTFS_DIR}/etc/nginx/sites-enabled/default"
